requirements.txt
Pillow==10.2.0
python-dateutil==2.8.2